pyproject.toml
[tool.poetry]
name = "AOframework"
version = "0.1.0"
description = "Web Automation Testing Framework With Playwright"
authors = ["AntonMelentiev <info@AntonMelentiev.com>"]

[tool.poetry.dependencies]
python = "^3.10"
allure-pytest = "^2.12.0"
pytest = "^7.2.1"
pytest-playwright = "^0.3.2"
pydantic = "^1.10.7"


[tool.poetry.dev-dependencies]
black = "^23.1.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
line-length = 120
extend-exclude = '''
(
  .venv
  | .idea
)
'''